+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 Notable changes to this project.
 
-## [dev]
+## [2.19.0] - 2024-07-26 - hello CryptoAPI
 - fix `SignalsAPI.ticker_universe`
 - added `validationBmcMean` and `era.validationBmc` to diagnostics
 - added `CyptoAPI`